pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "Backend for Infra"
authors = ["Bas Wilson <friedcatps@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.96.0"
uvicorn = "^0.22.0"
pymongo = "^4.3.3"
load-dotenv = "^0.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
